@@ -1,6 +1,10 @@
 InQuotex.task_class = 'EventTaskx::EventTask'
 InQuotex.supplier_class = 'Supplierx::Supplier'
-InQuotex.show_task_path = 'event_taskx.event_task_path(EventTaskx::EventTask.find_by_id(r.task_id))'
-InQuotex.return_suppliers = "Supplierx::Supplier.where(:active => true).order('name')"
+#InQuotex.show_task_path = 'event_taskx.event_task_path(EventTaskx::EventTask.find_by_id(r.task_id))'
+#InQuotex.return_suppliers = "Supplierx::Supplier.where(:active => true).order('name')"
 InQuotex.project_class = 'ExtConstructionProjectx::Project'
 InQuotex.customer_class = 'Kustomerx::Customer'
+InQuotex.category_class = 'Commonx::MiscDefinition'  #parking
+InQuotex.sub_category_class = 'Commonx::MiscDefinition'  #parking
+InQuotex.mfg_class = 'Commonx::MiscDefinition' #parking
+InQuotex.product_class = 'Commonx::MiscDefinition' #parking
