@@ -1,3 +1,3 @@
 module BillOfMaterialx
-  VERSION = "3.2.12.03"
+  VERSION = "01"
 end
