@@ -13,11 +13,10 @@ Rails.application.routes.draw do
   mount BizWorkflowx::Engine => '/biz_wf'
   mount EventTaskx::Engine => '/task'
   
-  resource :session
   
-  root :to => "authentify::sessions#new"
-  match '/signin',  :to => 'authentify::sessions#new'
-  match '/signout', :to => 'authentify::sessions#destroy'
-  match '/user_menus', :to => 'user_menus#index'
-  match '/view_handler', :to => 'authentify::application#view_handler'
+  root :to => "authentify/sessions#new"
+  get '/signin',  :to => 'authentify/sessions#new'
+  get '/signout', :to => 'authentify/sessions#destroy'
+  get '/user_menus', :to => 'user_menus#index'
+  get '/view_handler', :to => 'authentify/application#view_handler'
 end
