@@ -27,6 +27,6 @@ class CreateEventTaskxEventTasks < ActiveRecord::Migration
     add_index :event_taskx_event_tasks, :resource_string
     add_index :event_taskx_event_tasks, :task_category
     add_index :event_taskx_event_tasks, :wfid
-    add_index :event_taskx_event_tasks, [:resource_id, :resource_string]
+    add_index :event_taskx_event_tasks, [:resource_id, :resource_string], :name => :event_task_id_string
   end
 end
